Monthly Rides Trends Data
Based on the ridership data, it has been observed that there is a seasonal pattern in the usage of Cyclistic bikes. During the summer season, when the weather is pleasant and the days are longer,  members and casual riders are more likely to choose to ride. Interestingly, more member riders are using Cyclistic bikes than casual users. This indicates that summer months present an excellent opportunity to convert casual riders into members.